Fixes of sonar violations sdc and refactor
Issue-ID: SDC-1484

+++ catalog-dao/src/main/java/org/openecomp/sdc/be/dao/utils/ImageResizeUtil.java
@@ -129,12 +129,12 @@ public final class ImageResizeUtil {
 		int targetWidth = width;
 		int targetHeight = height;
 
-		float targetDimensions = Float.valueOf(width).floatValue() / Float.valueOf(height).floatValue();
-		float sourceDimensions = Float.valueOf(originalWidth).floatValue() / Float.valueOf(originalHeight).floatValue();
+		float targetDimensions = (float) width / (float) height;
+		float sourceDimensions = (float) originalWidth / (float) originalHeight;
 		if (targetDimensions > sourceDimensions) {
-			targetWidth = Float.valueOf(width * sourceDimensions / targetDimensions).intValue();
+			targetWidth = (int) (width * sourceDimensions / targetDimensions);
 		} else {
-			targetHeight = Float.valueOf(height * targetDimensions / sourceDimensions).intValue();
+			targetHeight = (int) (height * targetDimensions / sourceDimensions);
 		}
 
 		return new int[] { targetWidth, targetHeight };
